web/colores_favoritos/chall/core_backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
urllib3==1.26.12
